chg: Move Activity page to info/activiti.tsx
